requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
colorama==0.4.6
Cython==0.29.32
decorator==5.1.1
demucs3==3.0.1a0
diffq==0.2.3
dora-search==0.1.11
ffmpeg-python==0.2.0
future==0.18.2
idna==3.4
importlib-metadata==5.2.0
joblib==1.2.0
julius==0.2.7
lameenc==1.4.2
librosa==0.9.2
llvmlite==0.39.1
numba==0.56.4
numpy==1.23.5
omegaconf==2.3.0
opencv-python==4.6.0.66
openunmix==1.2.1
packaging==22.0
playsound==1.3.0
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
python-vlc==3.0.18121
PyYAML==6.0
requests==2.28.1
resampy==0.4.2
retrying==1.3.4
scikit-learn==1.2.0
scipy==1.9.3
simpleaudio==1.0.4
six==1.16.0
soundfile==0.11.0
SpeechRecognition==3.9.0
submitit==1.4.5
threadpoolctl==3.1.0
tk==0.1.0
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1
treetable==0.2.5
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
